% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_image.R
\name{get_species_image}
\alias{get_species_image}
\title{Download species image}
\usage{
get_species_image(species_name, image_number = 1)
}
\arguments{
\item{species_name}{Character string of species name.}

\item{image_number}{Single integer value or a vector of integers with the image number.}
}
\value{
A magick image file.
}
\description{
Download an image or series of images for a species, from the Reef Life Survey website.
}
\examples{
\dontrun{
# get a single image for the species
get_species_image("Pseudocheilinus hexataenia")
get_species_image("Pseudocheilinus hexataenia", 1)

# ...or a different single image
get_species_image("Pseudocheilinus hexataenia", 2)

# or get a series of images for the species (will ignore when not enough images are available)
get_species_image("Pseudocheilinus hexataenia", 1:30)

# NULL returns when no image available
get_species_image("Nonexistent species", 1)

# including the image in a ggplot
my_img <- get_species_image("Pseudocheilinus hexataenia", 1)

ggplot() +
  aes(x = 1:10,
      y = 1:10) +
  geom_point() +
  {if(!is.null(my_img))  annotation_raster(my_img,
                                           xmin = 1,
                                           xmax = 3,
                                           ymin = 8,
                                           ymax = 10) }
}

}
